+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 
+### 1.4.4
+- Do not require a theorem to be `Admitted.` for coqpilot to prove holes in it.
+- Correctly parse theorems that are declared with `Definition` keyword.
+
 ### 1.4.3
 - Tiny patch with shuffling of the hole array.